Make client register and sign test into integration test
