+++ crates/testing-utils/src/lib.rs
@@ -21,6 +21,8 @@ pub mod constants;
 pub mod create_test_keyshares;
 mod node_proc;
 pub mod substrate_context;
-pub use entropy_tss::helpers::tests::spawn_testing_validators;
+pub use entropy_tss::helpers::tests::{
+    jump_start_network_with_signer as jump_start_network, spawn_testing_validators,
+};
 pub use node_proc::TestNodeProcess;
 pub use substrate_context::*;
